Fix NavigationBar items NavigationBar items into NavigationBar from each Page main comoponents.
