+++ src/containers/EaglesEyeMain.tsx
@@ -4,14 +4,13 @@ import CountBullButtons from '@/components/CountBullButtons'
 import EaglesEyeBoard from '@/components/EaglesEyeBoard'
 import NavigationBar from '@/components/NavigationBar'
 import RoundBullScore from '@/components/RoundBullScore'
-import useLocale from '@/hooks/locale'
 import EaglesEyeGame from '@/lib/EaglesEyeGame/EaglesEyeGame'
 
 const EaglesEyeMain: FC = () => {
   const [game, setGame] = useState(new EaglesEyeGame())
   return (
     <div data-cy='eagles-eye-main'>
-      <NavigationBar items={GetNavItem()} />
+      <NavigationBar />
       <Grid templateRows={'repeat(3, auto)'} gap={10} justifyContent='center' p={10}>
         <GridItem>
           <CountBullButtons
@@ -44,41 +43,6 @@ const EaglesEyeMain: FC = () => {
       </Grid>
     </div>
   )
-}
-
-const GetNavItem = () => {
-  const { t } = useLocale()
-
-  return [
-    {
-      label: 'Games',
-      children: [
-        {
-          label: '501',
-          subLabel: t.N01_DESCRIPTION,
-          href: '/n01',
-        },
-        {
-          label: "Eagle's Eye",
-          subLabel: t.EAGLES_EYE_DESCRIPTION,
-          href: '#',
-        },
-        {
-          label: 'Cricket Number Count',
-          subLabel: t.CRICKET_NUMBER_COUNT_SHORT_DESCRIPTION,
-          href: '/cricketnumbercount',
-        },
-      ],
-    },
-    {
-      label: 'Respect',
-      href: '/respect',
-    },
-    {
-      label: 'History',
-      href: '/history',
-    },
-  ]
 }
 
 export default EaglesEyeMain
