+++ src/containers/N01main.tsx
@@ -6,7 +6,6 @@ import NavigationBar from '@/components/NavigationBar'
 import Players from '@/components/Players'
 import RoundScore from '@/components/RoundScore'
 import TargetBoard from '@/components/TargetBoard'
-import useLocale from '@/hooks/locale'
 import N01Game from '@/lib/N01Game/N01Game'
 import Player from '@/lib/Player/Player'
 
@@ -14,7 +13,7 @@ const N01main: FC = () => {
   const [game, setGame] = useState(new N01Game(501))
   return (
     <div data-cy='n01-main'>
-      <NavigationBar items={GetNavItem()} />
+      <NavigationBar />
       <Grid templateColumns='repeat(2, auto)' gap={10} p={4}>
         <GridItem>
           <Grid templateRows='repeat(3, auto)' gap={10}>
@@ -71,41 +70,6 @@ const N01main: FC = () => {
       </Grid>
     </div>
   )
-}
-
-const GetNavItem = () => {
-  const { t } = useLocale()
-
-  return [
-    {
-      label: 'Games',
-      children: [
-        {
-          label: '501',
-          subLabel: t.N01_DESCRIPTION,
-          href: '#',
-        },
-        {
-          label: "Eagle's Eye",
-          subLabel: t.EAGLES_EYE_DESCRIPTION,
-          href: '/eagleseye',
-        },
-        {
-          label: 'Cricket Number Count',
-          subLabel: t.CRICKET_NUMBER_COUNT_SHORT_DESCRIPTION,
-          href: '/cricketnumbercount',
-        },
-      ],
-    },
-    {
-      label: 'Respect',
-      href: '/respect',
-    },
-    {
-      label: 'History',
-      href: '/history',
-    },
-  ]
 }
 
 export default N01main
